+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `description` to `ArcGISWebMapElement.allowedDrawingTools.graphicAttributeOptions
+- `description` to `ArcGISWebMapElement.allowedDrawingTools.graphicAttributeOptions`
 
 ## [11.4.0] - 2025-06-03