+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `description` to `ArcGISWebMapElement.allowedDrawingTools.graphicAttributeOptions`
+
 ## [11.4.0] - 2025-06-03
 
 ### Added